run test cases for cacheUnaffected
