+++ test/WatchCacheUnaffectedTestCases.longtest.js
@@ -1,0 +1,11 @@
+const { describeCases } = require("./WatchTestCases.template");
+
+describeCases({
+	name: "WatchCacheUnaffectedTestCases",
+	experiments: {
+		cacheUnaffected: true
+	},
+	optimization: {
+		sideEffects: false
+	}
+});
