+++ test/WatchTestCases.longtest.js
@@ -1,0 +1,5 @@
+const { describeCases } = require("./WatchTestCases.template");
+
+describeCases({
+	name: "WatchTestCases"
+});
