+++ test/watchCases/cache/add-defines/test.filter.js
@@ -1,0 +1,3 @@
+module.exports = function (config) {
+	return !(config.experiments && config.experiments.cacheUnaffected);
+};
